FUSETOOLS-2710 - factorize duplicated code in syndesis integration test
- extract a common class in test.util plugin
- removed several unused plugin dependencies
- still several part of code with very little modifications which would
be best to factorize
