+++ editor/tests/org.fusesource.ide.projecttemplates.tests.integration/META-INF/MANIFEST.MF
@@ -31,8 +31,8 @@ Require-Bundle: org.junit,
  org.fusesource.ide.foundation.ui,
  org.fusesource.ide.camel.model.service.core,
  org.fusesource.ide.camel.model.service.impl,
- org.fusesource.ide.camel.editor;bundle-version="10.0.0",
  org.fusesource.ide.launcher;bundle-version="10.0.0",
+ org.fusesource.ide.camel.editor;bundle-version="10.0.0",
  org.fusesource.ide.launcher.ui,
  org.fusesource.ide.projecttemplates;bundle-version="10.0.0",
  org.fusesource.ide.project;bundle-version="10.0.0",
