+++ syndesis/tests/org.fusesource.ide.syndesis.extension.ui.tests.integration/META-INF/MANIFEST.MF
@@ -8,10 +8,12 @@ Require-Bundle: org.junit;bundle-version="4.12.0",
  org.assertj.core;bundle-version="2.1.0",
  org.eclipse.debug.ui;bundle-version="3.12.50",
  org.eclipse.wst.common.project.facet.ui;bundle-version="1.4.600",
- org.eclipse.m2e.core;bundle-version="1.8.2",
- org.eclipse.m2e.maven.runtime;bundle-version="1.8.2",
+ org.eclipse.m2e.core;bundle-version="1.6.2",
+ org.eclipse.m2e.maven.runtime;bundle-version="1.6.2",
  org.jboss.tools.foundation.core;bundle-version="1.3.1",
  org.jboss.tools.foundation.ui;bundle-version="1.3.1",
+ org.eclipse.graphiti;bundle-version="0.14.0",
+ org.eclipse.graphiti.ui;bundle-version="0.14.0",
  org.fusesource.ide.preferences,
  org.fusesource.ide.foundation.core,
  org.fusesource.ide.foundation.ui,
